ml_worker/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
onnxruntime==1.20.1
av==14.0.1
Pillow==11.0.0
numpy==2.1.2
huggingface-hub==0.26.2
